.history/requirements_20250726173436.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
fake-useragent==1.4.0
pandas==2.1.3
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
colorama==0.4.6
tqdm==4.66.1
